refactor: api converters
