+++ packages/bff/src/model/api/apiConverter.ts
@@ -18,8 +18,10 @@ import {
   authorizationApi,
   bffApi,
   catalogApi,
+  selfcareV2ClientApi,
   tenantApi,
 } from "pagopa-interop-api-clients";
+import { match, P } from "ts-pattern";
 import { agreementApiState, catalogApiDescriptorState } from "./apiTypes.js";
 
 export function toDescriptorWithOnlyAttributes(
@@ -169,16 +171,26 @@ export function toTenantWithOnlyAttributes(
   };
 }
 
-export function toAuthorizationKeySeed(
-  seed: bffApi.KeySeed,
+export const toBffApiCompactClient = (
+  input: authorizationApi.ClientWithKeys
+): bffApi.CompactClient => ({
+  hasKeys: input.keys.length > 0,
+  id: input.client.id,
+  name: input.client.name,
+});
+
+export const toBffApiCompactUser = (
+  input: selfcareV2ClientApi.UserResponse,
   userId: string
-): authorizationApi.KeySeed {
-  return {
-    userId,
-    key: seed.key,
-    use: seed.use,
-    alg: seed.alg,
-    name: seed.name,
-    createdAt: new Date().toISOString(),
-  };
-}
+): bffApi.CompactUser =>
+  match(input)
+    .with({ name: P.nullish, surname: P.nullish }, () => ({
+      userId,
+      name: "Utente",
+      familyName: userId,
+    }))
+    .otherwise((ur) => ({
+      userId,
+      name: ur.name ?? "",
+      familyName: ur.surname ?? "",
+    }));
